// src/app/model/license-plate.ts
export class LicensePlate {
   id: number | undefined;
   value: string | undefined;
   userId: number | undefined;
   place: string | undefined;
   number: string | undefined;
   code: string | undefined;
   creationDate: Date | undefined;
   expirationDate: Date | undefined;
   type: string | undefined;
}